+++ packages/pyroscope-panel-plugin/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.5](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/panel-plugin@1.4.4...@pyroscope/panel-plugin@1.4.5) (2022-11-17)
+
+**Note:** Version bump only for package @pyroscope/panel-plugin
+
+
+
+
+
+## [1.4.4](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/panel-plugin@1.4.3...@pyroscope/panel-plugin@1.4.4) (2022-11-17)
+
+**Note:** Version bump only for package @pyroscope/panel-plugin
+
+
+
+
+
 ## [1.4.3](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/panel-plugin@1.4.2...@pyroscope/panel-plugin@1.4.3) (2022-11-15)
 
 **Note:** Version bump only for package @pyroscope/panel-plugin